Moved example bot implementation to separate module
Also performed package cleanup
