+++ .idea/artifacts/Sholimp.xml
@@ -1,10 +1,11 @@
 <component name="ArtifactManager">
-  <artifact type="jar" name="VimeBot">
+  <artifact type="jar" name="Sholimp">
     <output-path>$PROJECT_DIR$/out/</output-path>
-    <root id="archive" name="VimeBot.jar">
-      <element id="module-output" name="VimeBot" />
+    <root id="archive" name="Sholimp.jar">
+      <element id="module-output" name="Sholimp" />
       <element id="extracted-dir" path="$PROJECT_DIR$/lib/jna-5.3.1.jar" path-in-jar="/" />
       <element id="extracted-dir" path="$PROJECT_DIR$/lib/jna-platform-5.3.1.jar" path-in-jar="/" />
+      <element id="module-output" name="VimeBot" />
     </root>
   </artifact>
 </component>
