+++ VimeBot/src/pro/delfik/vimebot/command/Command.java
@@ -1,4 +1,4 @@
-package pro.delfik.vimebot;
+package pro.delfik.vimebot.command;
 
 import lombok.Getter;
