+++ VimeBot/src/pro/delfik/vimebot/command/CommandPattern.java
@@ -1,4 +1,4 @@
-package pro.delfik.vimebot;
+package pro.delfik.vimebot.command;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
